sysmobts: Fix the function check_oml_msg of utils
This patch allows check the oml message correctly.
I have splitted the function in three different, one for
checking the manufacturer label ID, another for checking the
FOM and the principal which will call to the other function
for checking the OM header and all the part that I have told
before of the OML message received.

Signed-off-by: Alvaro Neira Ayuso <anayuso@sysmocom.de>