A dedicated OML router program (osmobts-omlrouter)
The idea of this OML router is to be the entity that connects
the OML link to the BSC.  osmo-bts as well as other programs like
sysmobts-mgr and possibly more will then connect to the OML-router
rather than the BSC.

The point is that those "OML clients" can then register for certain
OML messages (particularly the managed objects like TRX, etc.) which
then get routed to them.

This is particularly useful in the context of 'stacked' multi-TRX
BTSs, where we will have two osmo-bts processes, one on each of the
TRX, both connecting OML to the oml-router.  Through their respective
subscription of the baseband transceiver OML object, they will get
routed the IPA CONNECT RSL message from the BSC, and both osmo-bts
processes will establish independent RSL connections to the BSC.