power/sysmobts: Add a manual ms power level control
Currently the DSP is instructed to achieve a given uplink
power target but there are circumstances (e.g. EMV testing)
where we need more control over it. The "manual/software/osmo"
power control can only be implemented per TRX and not per
lchan. Add a very very basic control that checks the MS Power
used by the phone, the actual receive level and then adjust
the power.

The code doesn't take the history into account, if the phone
can not reach the requested power level the code will be stuck
(e.g. no timeout based on multiframes). It has a mode for a
fixed power control but no way to set it yet.

The change of the mode requires a restart of the software.

Conflicts:
	include/osmo-bts/bts_model.h
	src/common/vty.c
	src/osmo-bts-sysmo/l1_if.c
	src/osmo-bts-sysmo/l1_if.h
	src/osmo-bts-sysmo/oml.c
	tests/sysmobts/sysmobts_test.c