Add code to control the PA of the SOB-BTS
At least in v1 of the SOB-BTS, there is a SX1502 I2C GPIO expander
to control the two different PAs in the system.  This adds a SX150x
driver on top of Linux userspae i2c-dev access, and a thin layer of API
functions as well as a command line tool.

The code is currently still untested.