vty: Add commands to manually activate/deactivate a channel.
This is a hack, but it's the easiest way to make BTS level loopback to work.
Another way to implement this is to have BSC/NITB to send the OML command.

Note, that the current code hardcode the channel to be TCH/F which is what we
need for the BER testing.